Atualizacao de backup

+++ app/src/main/java/com/example/fasipemobilej/model/response/UserResponse.java
@@ -2,7 +2,7 @@ package com.example.fasipemobilej.model.response;
 
 public record UserResponse(
         String nome,
-        Long tipo,
+        int tipo,
         Long supervisor,
         int status,
         String consProf,
